+++ src/main/java/io/bookwright/junit/PreconditionProvider.java
@@ -25,6 +25,7 @@ public class PreconditionProvider implements BeforeTestExecutionCallback {
 
         ApiSteps api = StepsParameterResolver.injectorFor(ApiSteps.class, context).getInstance(ApiSteps.class);
         ExtensionContext.Store store = NamespaceRegistry.methodStore(context);
+        TestDataExtension.getOrCreate(context);
 
         for (IPrecondition precondition : preconditions) {
             Allure.step("Precondition: " + precondition.title(), () -> precondition.execute(api, store));